Migrate AboutActivity to fragment
